Memory footprint size reduction
This change reduces the persistant storage of about 3 copies of the
tiff image as cv::Mat objects.

It makes temporary FrameData::frameF, FrameData::frame_bgoff. These
changes make no functional change to the program.

It also removes FrameData::frame_signalU, to use only
FrameData::frame_signal. The former was used to display the signal in
the signal window. Use of the CV_8UC3 format for the data in
frame_signalU meant that text and lines superimposed on the signal
window were able to be antialiased (CV_8UC3 has an alpha
channel). However, in the interest of reducing memory use, I will just
use frame_signal (CV_32FC3 format) for the signal window, and not
worry that the lines and text will be a little blocky.